refactor converters to make room for more in the future
