+++ packages/rest/src/runtime/shared/index.mts
@@ -4,7 +4,7 @@ export {
     UnparsedHttpEvent, SerializedHttpResponse
 } from './http-event.mjs';
 export {NornirRestRequestError, NornirRestError, httpErrorHandler, mapError, mapErrorClass, NornirRouteNotFoundError, NornirRestRequestValidationError, ErrorMapping} from './error.mjs'
-export {ApiGatewayProxyV2, Express, startLocalServer} from "./converters.mjs"
+export {ApiGatewayProxyV2, Express, startLocalServer, Converter} from "./converters/index.mjs"
 export {httpEventParser, HttpBodyParser, HttpBodyParserMap, HttpQueryStringParser, NornirRestParseError} from "./parse.mjs"
 export {httpResponseSerializer, HttpBodySerializer, HttpBodySerializerMap} from "./serialize.mjs"
 export {normalizeEventHeaders, normalizeHeaders, getContentType} from "./utils.mjs"
